requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
ansitable==0.9.7
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.6
colored==1.4.4
commentjson==0.9.0
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.12.4
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.40
glcontext==2.5.0
google-auth==2.23.2
google-auth-oauthlib==1.0.0
grpcio==1.59.0
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
importlib-resources==6.1.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lark-parser==0.7.8
lazy_loader==0.3
lightning-utilities==0.9.0
lit==17.0.2
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
moderngl==5.8.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
open3d~=0.15.2
opencv-python==4.6.0.66
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.0.3
pathtools==0.1.2
Pillow==10.0.1
portalocker==2.8.2
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.7.1
pycolmap>=0.3.0,<=0.4.0
pyntcloud==0.1.5
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
pytorch-lightning==1.9.5
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@eaf0709d6af0025fe94d1ee7cec454bc3054826a
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.13.0
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
spatialmath-python==1.0.5
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.10
tqdm==4.66.1
trimesh==3.23.5
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.120
urllib3==2.0.6
wandb==0.13.11
Werkzeug==3.0.0
yacs==0.1.8
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0